+++ src/chat-sidebar.tsx
@@ -46,7 +46,6 @@ import {
   VscEyeClosed,
   VscTriangleRight,
   VscTriangleDown,
-  VscWarning,
   VscSettingsGear,
   VscPassFilled,
   VscTools,
@@ -56,14 +55,6 @@ import {
 import { MdOutlineCheckBoxOutlineBlank, MdCheckBox } from 'react-icons/md';
 
 import { extractLLMGeneratedCode, isDarkTheme } from './utils';
-import * as path from 'path';
-
-const OPENAI_COMPATIBLE_CHAT_MODEL_ID = 'openai-compatible-chat-model';
-const LITELLM_COMPATIBLE_CHAT_MODEL_ID = 'litellm-compatible-chat-model';
-const OPENAI_COMPATIBLE_INLINE_COMPLETION_MODEL_ID =
-  'openai-compatible-inline-completion-model';
-const LITELLM_COMPATIBLE_INLINE_COMPLETION_MODEL_ID =
-  'litellm-compatible-inline-completion-model';
 
 export enum RunChatCompletionType {
   Chat,
@@ -262,30 +253,6 @@ export class GitHubCopilotLoginDialogBody extends ReactWidget {
   }
 
   private _onLoggedIn: () => void;
-}
-
-export class SettingsDialogBody extends ReactWidget {
-  constructor(options: {
-    onSave: () => void;
-    onEditMCPConfigClicked: () => void;
-  }) {
-    super();
-
-    this._onEditMCPConfigClicked = options.onEditMCPConfigClicked;
-    this._onSave = options.onSave;
-  }
-
-  render(): JSX.Element {
-    return (
-      <SettingsDialogBodyComponent
-        onEditMCPConfigClicked={this._onEditMCPConfigClicked}
-        onSave={this._onSave}
-      />
-    );
-  }
-
-  private _onEditMCPConfigClicked: () => void;
-  private _onSave: () => void;
 }
 
 interface IChatMessageContent {
@@ -2494,530 +2461,6 @@ function GitHubCopilotLoginDialogBodyComponent(props: any) {
           </button>
         </div>
       )}
-    </div>
-  );
-}
-
-function SettingsDialogBodyComponent(props: any) {
-  const nbiConfig = NBIAPI.config;
-  const llmProviders = nbiConfig.llmProviders;
-  const [chatModels, setChatModels] = useState([]);
-  const [inlineCompletionModels, setInlineCompletionModels] = useState([]);
-  const [mcpServerNames, setMcpServerNames] = useState(
-    nbiConfig.toolConfig.mcpServers?.map((server: any) => server.id) || []
-  );
-
-  const handleSaveClick = async () => {
-    const config: any = {
-      default_chat_mode: defaultChatMode,
-      chat_model: {
-        provider: chatModelProvider,
-        model: chatModel,
-        properties: chatModelProperties
-      },
-      inline_completion_model: {
-        provider: inlineCompletionModelProvider,
-        model: inlineCompletionModel,
-        properties: inlineCompletionModelProperties
-      }
-    };
-
-    if (
-      chatModelProvider === 'github-copilot' ||
-      inlineCompletionModelProvider === 'github-copilot'
-    ) {
-      config.store_github_access_token = storeGitHubAccessToken;
-    }
-
-    await NBIAPI.setConfig(config);
-
-    props.onSave();
-  };
-
-  const handleRefreshOllamaModelListClick = async () => {
-    await NBIAPI.updateOllamaModelList();
-    updateModelOptionsForProvider(chatModelProvider, 'chat');
-  };
-
-  const [chatModelProvider, setChatModelProvider] = useState(
-    nbiConfig.chatModel.provider || 'none'
-  );
-  const [inlineCompletionModelProvider, setInlineCompletionModelProvider] =
-    useState(nbiConfig.inlineCompletionModel.provider || 'none');
-  const [defaultChatMode, setDefaultChatMode] = useState<string>(
-    nbiConfig.defaultChatMode
-  );
-  const [chatModel, setChatModel] = useState<string>(nbiConfig.chatModel.model);
-  const [chatModelProperties, setChatModelProperties] = useState<any[]>([]);
-  const [inlineCompletionModelProperties, setInlineCompletionModelProperties] =
-    useState<any[]>([]);
-  const [inlineCompletionModel, setInlineCompletionModel] = useState(
-    nbiConfig.inlineCompletionModel.model
-  );
-  const [storeGitHubAccessToken, setStoreGitHubAccessToken] = useState(
-    nbiConfig.storeGitHubAccessToken
-  );
-
-  const updateModelOptionsForProvider = (
-    providerId: string,
-    modelType: 'chat' | 'inline-completion'
-  ) => {
-    if (modelType === 'chat') {
-      setChatModelProvider(providerId);
-    } else {
-      setInlineCompletionModelProvider(providerId);
-    }
-    const models =
-      modelType === 'chat'
-        ? nbiConfig.chatModels
-        : nbiConfig.inlineCompletionModels;
-    const selectedModelId =
-      modelType === 'chat'
-        ? nbiConfig.chatModel.model
-        : nbiConfig.inlineCompletionModel.model;
-
-    const providerModels = models.filter(
-      (model: any) => model.provider === providerId
-    );
-    if (modelType === 'chat') {
-      setChatModels(providerModels);
-    } else {
-      setInlineCompletionModels(providerModels);
-    }
-    let selectedModel = providerModels.find(
-      (model: any) => model.id === selectedModelId
-    );
-    if (!selectedModel) {
-      selectedModel = providerModels?.[0];
-    }
-    if (selectedModel) {
-      if (modelType === 'chat') {
-        setChatModel(selectedModel.id);
-        setChatModelProperties(selectedModel.properties);
-      } else {
-        setInlineCompletionModel(selectedModel.id);
-        setInlineCompletionModelProperties(selectedModel.properties);
-      }
-    } else {
-      if (modelType === 'chat') {
-        setChatModelProperties([]);
-      } else {
-        setInlineCompletionModelProperties([]);
-      }
-    }
-  };
-
-  const onModelPropertyChange = (
-    modelType: 'chat' | 'inline-completion',
-    propertyId: string,
-    value: string
-  ) => {
-    const modelProperties =
-      modelType === 'chat'
-        ? chatModelProperties
-        : inlineCompletionModelProperties;
-    const updatedProperties = modelProperties.map((property: any) => {
-      if (property.id === propertyId) {
-        return { ...property, value };
-      }
-      return property;
-    });
-    if (modelType === 'chat') {
-      setChatModelProperties(updatedProperties);
-    } else {
-      setInlineCompletionModelProperties(updatedProperties);
-    }
-  };
-
-  const handleReloadMCPServersClick = async () => {
-    const data = await NBIAPI.reloadMCPServerList();
-    setMcpServerNames(data.mcpServers?.map((server: any) => server.id) || []);
-  };
-
-  useEffect(() => {
-    updateModelOptionsForProvider(chatModelProvider, 'chat');
-    updateModelOptionsForProvider(
-      inlineCompletionModelProvider,
-      'inline-completion'
-    );
-  }, []);
-
-  return (
-    <div className="config-dialog">
-      <div className="config-dialog-body">
-        <div className="model-config-section">
-          <div className="model-config-section-header">Default chat mode</div>
-          <div className="model-config-section-body">
-            <div className="model-config-section-row">
-              <div className="model-config-section-column">
-                <div>
-                  <select
-                    className="jp-mod-styled"
-                    value={defaultChatMode}
-                    onChange={event => setDefaultChatMode(event.target.value)}
-                  >
-                    <option value="ask">Ask</option>
-                    <option value="agent">Agent</option>
-                  </select>
-                </div>
-              </div>
-              <div className="model-config-section-column"> </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="model-config-section">
-          <div className="model-config-section-header">Chat model</div>
-          <div className="model-config-section-body">
-            <div className="model-config-section-row">
-              <div className="model-config-section-column">
-                <div>Provider</div>
-                <div>
-                  <select
-                    className="jp-mod-styled"
-                    onChange={event =>
-                      updateModelOptionsForProvider(event.target.value, 'chat')
-                    }
-                  >
-                    {llmProviders.map((provider: any, index: number) => (
-                      <option
-                        key={index}
-                        value={provider.id}
-                        selected={provider.id === chatModelProvider}
-                      >
-                        {provider.name}
-                      </option>
-                    ))}
-                    <option
-                      key={-1}
-                      value="none"
-                      selected={
-                        chatModelProvider === 'none' ||
-                        !llmProviders.find(
-                          provider => provider.id === chatModelProvider
-                        )
-                      }
-                    >
-                      None
-                    </option>
-                  </select>
-                </div>
-              </div>
-              {!['openai-compatible', 'litellm-compatible', 'none'].includes(
-                chatModelProvider
-              ) &&
-                chatModels.length > 0 && (
-                  <div className="model-config-section-column">
-                    <div>Model</div>
-                    {![
-                      OPENAI_COMPATIBLE_CHAT_MODEL_ID,
-                      LITELLM_COMPATIBLE_CHAT_MODEL_ID
-                    ].includes(chatModel) &&
-                      chatModels.length > 0 && (
-                        <div>
-                          <select
-                            className="jp-mod-styled"
-                            onChange={event => setChatModel(event.target.value)}
-                          >
-                            {chatModels.map((model: any, index: number) => (
-                              <option
-                                key={index}
-                                value={model.id}
-                                selected={model.id === chatModel}
-                              >
-                                {model.name}
-                              </option>
-                            ))}
-                          </select>
-                        </div>
-                      )}
-                  </div>
-                )}
-            </div>
-
-            <div className="model-config-section-row">
-              <div className="model-config-section-column">
-                {chatModelProvider === 'ollama' && chatModels.length === 0 && (
-                  <div className="ollama-warning-message">
-                    No Ollama models found! Make sure{' '}
-                    <a href="https://ollama.com/" target="_blank">
-                      Ollama
-                    </a>{' '}
-                    is running and models are downloaded to your computer.{' '}
-                    <a
-                      href="javascript:void(0)"
-                      onClick={handleRefreshOllamaModelListClick}
-                    >
-                      Try again
-                    </a>{' '}
-                    once ready.
-                  </div>
-                )}
-              </div>
-            </div>
-
-            <div className="model-config-section-row">
-              <div className="model-config-section-column">
-                {chatModelProperties.map((property: any, index: number) => (
-                  <div className="form-field-row" key={index}>
-                    <div className="form-field-description">
-                      {property.name} {property.optional ? '(optional)' : ''}
-                    </div>
-                    <input
-                      name="chat-model-id-input"
-                      placeholder={property.description}
-                      className="jp-mod-styled"
-                      spellCheck={false}
-                      value={property.value}
-                      onChange={event =>
-                        onModelPropertyChange(
-                          'chat',
-                          property.id,
-                          event.target.value
-                        )
-                      }
-                    />
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="model-config-section">
-          <div className="model-config-section-header">Auto-complete model</div>
-          <div className="model-config-section-body">
-            <div className="model-config-section-row">
-              <div className="model-config-section-column">
-                <div>Provider</div>
-                <div>
-                  <select
-                    className="jp-mod-styled"
-                    onChange={event =>
-                      updateModelOptionsForProvider(
-                        event.target.value,
-                        'inline-completion'
-                      )
-                    }
-                  >
-                    {llmProviders.map((provider: any, index: number) => (
-                      <option
-                        key={index}
-                        value={provider.id}
-                        selected={provider.id === inlineCompletionModelProvider}
-                      >
-                        {provider.name}
-                      </option>
-                    ))}
-                    <option
-                      key={-1}
-                      value="none"
-                      selected={
-                        inlineCompletionModelProvider === 'none' ||
-                        !llmProviders.find(
-                          provider =>
-                            provider.id === inlineCompletionModelProvider
-                        )
-                      }
-                    >
-                      None
-                    </option>
-                  </select>
-                </div>
-              </div>
-              {!['openai-compatible', 'litellm-compatible', 'none'].includes(
-                inlineCompletionModelProvider
-              ) && (
-                <div className="model-config-section-column">
-                  <div>Model</div>
-                  {![
-                    OPENAI_COMPATIBLE_INLINE_COMPLETION_MODEL_ID,
-                    LITELLM_COMPATIBLE_INLINE_COMPLETION_MODEL_ID
-                  ].includes(inlineCompletionModel) && (
-                    <div>
-                      <select
-                        className="jp-mod-styled"
-                        onChange={event =>
-                          setInlineCompletionModel(event.target.value)
-                        }
-                      >
-                        {inlineCompletionModels.map(
-                          (model: any, index: number) => (
-                            <option
-                              key={index}
-                              value={model.id}
-                              selected={model.id === inlineCompletionModel}
-                            >
-                              {model.name}
-                            </option>
-                          )
-                        )}
-                      </select>
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-
-            <div className="model-config-section-row">
-              <div className="model-config-section-column">
-                {inlineCompletionModelProperties.map(
-                  (property: any, index: number) => (
-                    <div className="form-field-row" key={index}>
-                      <div className="form-field-description">
-                        {property.name} {property.optional ? '(optional)' : ''}
-                      </div>
-                      <input
-                        name="inline-completion-model-id-input"
-                        placeholder={property.description}
-                        className="jp-mod-styled"
-                        spellCheck={false}
-                        value={property.value}
-                        onChange={event =>
-                          onModelPropertyChange(
-                            'inline-completion',
-                            property.id,
-                            event.target.value
-                          )
-                        }
-                      />
-                    </div>
-                  )
-                )}
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {(chatModelProvider === 'github-copilot' ||
-          inlineCompletionModelProvider === 'github-copilot') && (
-          <div className="model-config-section">
-            <div className="model-config-section-header access-token-config-header">
-              GitHub Copilot login{' '}
-              <a
-                href="https://github.com/notebook-intelligence/notebook-intelligence/blob/main/README.md#remembering-github-copilot-login"
-                target="_blank"
-              >
-                {' '}
-                <VscWarning
-                  className="access-token-warning"
-                  title="Click to learn more about security implications"
-                />
-              </a>
-            </div>
-            <div className="model-config-section-body">
-              <div className="model-config-section-row">
-                <div className="model-config-section-column">
-                  <label>
-                    <input
-                      type="checkbox"
-                      checked={storeGitHubAccessToken}
-                      onChange={event => {
-                        setStoreGitHubAccessToken(event.target.checked);
-                      }}
-                    />
-                    Remember my GitHub Copilot access token
-                  </label>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-
-        <div className="model-config-section">
-          <div className="model-config-section-header">
-            MCP Servers ({mcpServerNames.length}) [
-            <a href="javascript:void(0)" onClick={props.onEditMCPConfigClicked}>
-              edit
-            </a>
-            ]
-          </div>
-          <div className="model-config-section-body">
-            <div className="model-config-section-row">
-              <div className="model-config-section-column">
-                {mcpServerNames.length === 0 && (
-                  <div>
-                    No MCP servers found. Add MCP servers in the configuration
-                    file.
-                  </div>
-                )}
-                {mcpServerNames.length > 0 && (
-                  <div>{mcpServerNames.sort().join(', ')}</div>
-                )}
-              </div>
-              <div
-                className="model-config-section-column"
-                style={{ flexGrow: 'initial' }}
-              >
-                <button
-                  className="jp-Dialog-button jp-mod-reject jp-mod-styled"
-                  onClick={handleReloadMCPServersClick}
-                >
-                  <div className="jp-Dialog-buttonLabel">Reload</div>
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div className="model-config-section">
-          <div className="model-config-section-header">Config file path</div>
-          <div className="model-config-section-body">
-            <div className="model-config-section-row">
-              <div className="model-config-section-column">
-                <span
-                  className="user-code-span"
-                  onClick={() => {
-                    navigator.clipboard.writeText(
-                      path.join(NBIAPI.config.userConfigDir, 'config.json')
-                    );
-                    return true;
-                  }}
-                >
-                  {path.join(NBIAPI.config.userConfigDir, 'config.json')}{' '}
-                  <span
-                    className="copy-icon"
-                    dangerouslySetInnerHTML={{ __html: copySvgstr }}
-                  ></span>
-                </span>
-              </div>
-            </div>
-          </div>
-          <div className="model-config-section-header">
-            MCP config file path
-          </div>
-          <div className="model-config-section-body">
-            <div className="model-config-section-row">
-              <div className="model-config-section-column">
-                <span
-                  className="user-code-span"
-                  onClick={() => {
-                    navigator.clipboard.writeText(
-                      path.join(NBIAPI.config.userConfigDir, 'mcp.json')
-                    );
-                    return true;
-                  }}
-                >
-                  {path.join(NBIAPI.config.userConfigDir, 'mcp.json')}{' '}
-                  <span
-                    className="copy-icon"
-                    dangerouslySetInnerHTML={{ __html: copySvgstr }}
-                  ></span>
-                </span>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div className="config-dialog-footer">
-        <button
-          className="jp-Dialog-button jp-mod-accept jp-mod-styled"
-          onClick={handleSaveClick}
-        >
-          <div className="jp-Dialog-buttonLabel">Save</div>
-        </button>
-      </div>
     </div>
   );
 }
